+++ xcommon.cmake
@@ -32,8 +32,8 @@ set_property(GLOBAL PROPERTY SSH_HOST_FAILURE "")
 
 set(MANIFEST_OUT ${CMAKE_BINARY_DIR}/manifest.txt)
 if(FULL_MANIFEST)
-    set(DEP_REQ_HEADER "                                  | Dependency_requirement")
-    set(DEP_REQ_DIVIDER "+-----------------------------------------")
+    set(DEP_REQ_HEADER "                                  | Dependency_requirement | Depends_on")
+    set(DEP_REQ_DIVIDER                                  "+------------------------+------------------------------")
 endif()
 set(MANIFEST_HEADER
         "Name                    | Location                                        | Branch/tag             | Changeset${DEP_REQ_HEADER}\n"
@@ -368,7 +368,7 @@ endfunction()
 # Called for the top-level app/lib and each module, this takes a name and version (either blank for the
 # top-level app or provided by the module dependency specification), checks the git repo status to work out
 # changeset hashes and tags, and then appends an entry in the manifest file for the module.
-function(manifest_git_status name version)
+function(manifest_git_status name version manifest_str_ret)
     if(NOT name)
         set(working_dir ${CMAKE_SOURCE_DIR})
     else()
@@ -415,7 +415,9 @@ function(manifest_git_status name version)
     endif()
 
     form_manifest_string("${name}" "${version}" "${repo}" "${branch}" "${tag}" "${commit_hash}" "${status}" manifest_str)
-    file(APPEND ${MANIFEST_OUT} "${manifest_str}\n")
+
+    set(${manifest_str_ret} ${manifest_str} PARENT_SCOPE)
+
 endfunction()
 
 ## Registers an application and its dependencies
@@ -562,7 +564,19 @@ function(XMOS_REGISTER_APP)
 
     # Overwrites file if already present and then record manifest entry for application repo
     file(WRITE ${MANIFEST_OUT} ${MANIFEST_HEADER})
-    manifest_git_status("" "")
+    manifest_git_status("" "" manifest_str)
+
+    # Create the Depends_on column in manifest
+    if(${FULL_MANIFEST})
+        pad_string(${manifest_str} 170 manifest_str)
+        set(DEP_NAME_LIST "")
+        foreach(DEP_MODULE ${LIB_DEPENDENT_MODULES})
+            parse_dep_string(${DEP_MODULE} DEP_REPO DEP_VERSION DEP_NAME)
+            string(APPEND DEP_NAME_LIST "${DEP_NAME} ")
+        endforeach()
+    endif()
+    file(APPEND ${MANIFEST_OUT} ${manifest_str})
+    file(APPEND ${MANIFEST_OUT} "${DEP_NAME_LIST}\n")
 
     set(current_module ${PROJECT_NAME})
     XMOS_REGISTER_DEPS()
@@ -760,9 +774,23 @@ function(XMOS_REGISTER_DEPS)
             message(STATUS "Adding dependency ${DEP_NAME}")
             include(${module_dir}/lib_build_info.cmake)
 
-            manifest_git_status(${DEP_NAME} ${DEP_VERSION})
+            manifest_git_status(${DEP_NAME} ${DEP_VERSION} manifest_str)
+
+            # Create the Depends_on column in manifest
+            if(${FULL_MANIFEST})
+                pad_string(${manifest_str} 170 manifest_str)
+                set(DEP_NAME_LIST "")
+                foreach(DEP_MODULE ${LIB_DEPENDENT_MODULES})
+                    parse_dep_string(${DEP_MODULE} DEP_REPO DEP_VERSION DEP_NAME)
+                    string(APPEND DEP_NAME_LIST "${DEP_NAME} ")
+                endforeach()
+            endif()
+            file(APPEND ${MANIFEST_OUT} ${manifest_str})
+            file(APPEND ${MANIFEST_OUT} "${DEP_NAME_LIST}\n")
         endif()
     endforeach()
+
+
 endfunction()
 
 ## Registers a static library target
@@ -826,7 +854,8 @@ function(XMOS_STATIC_LIBRARY)
 
     # Overwrites file if already present and then record manifest entry for application repo
     file(WRITE ${MANIFEST_OUT} ${MANIFEST_HEADER})
-    manifest_git_status("" "")
+    manifest_git_status("" "" manifest_str)
+    file(APPEND ${MANIFEST_OUT} "${manifest_str}\n")
 
     set(current_module ${LIB_NAME})
     XMOS_REGISTER_DEPS()